Model interception
+ abc can now intercept model runs which exceeded already the given
epsilon limit
+ tumor growth model compares to data on the fly

+++ src/tumor3d/Montecarlo.h
@@ -65,7 +65,7 @@ extern int Case;
 
 
 #ifndef __WithGUI__
-	int montecarlo( int argc, char **argv);
+	double montecarlo( int argc, char **argv);
 #else
 	#include <gui/window.h>
 	int montecarlo( int argc, char **argv, Window *gui);
